Expense importing from csv, prod fix Duplicate check too sensitive.

+++ application/views/expenses/import/captured.php
@@ -8,7 +8,7 @@
                 ?>
                 <table id="expenseSummary" class="tablesorter responsive expanded widget-zebra">
                     <thead>
-                    <th/>
+                    <th>Status</th>
                     <th>Date</th>
                     <th>Expense Type</th>
                     <th>Payment Method</th>
@@ -22,20 +22,25 @@
                         $total = 0.0;
                         foreach ($expenses as $k => $v) {
                             echo "<tr>";
-                            echo "<td>" . ++$k . "</td>";
+                            echo "<td>" . $v["status"] . ": ".$v["statusMessage"]."</td>";
                             echo "<td>" . $v["expense_date"] . "</td>";
                             echo "<td>" . $expenseTypes[$v["expense_type_id"]]["description"] . "</td>";
                             echo "<td>" . $expensePaymentMethod[$v["payment_method_id"]]["description"] . "</td>";
                             echo "<td>" . $v["description"] . "</td>";
                             echo "<td>" . $v["location"] . "</td>";
                             echo "<td class='align-right'>" . $v["amount"] . "</td>";
-                            echo "<td><a href='/expenses/edit/" . $v["id"] . "'>Edit</a>&nbsp;&nbsp;|&nbsp;&nbsp;<a href='/expenses/delete/" . $v["id"] . "' onclick='return confirm_delete()'>Delete</a></td>";
+                            if($v["status"] == "Success"){
+                                echo "<td><a href='/expenses/edit/" . $v["id"] . "'>Edit</a>&nbsp;&nbsp;|&nbsp;&nbsp;<a href='/expenses/delete/" . $v["id"] . "' onclick='return confirm_delete()'>Delete</a></td>";
+                            }else{
+                                echo "<td>&nbsp;</td>";
+                            }
                             echo "</tr>";
                             $total += $v["amount"];
                         }
                         echo "<tr class='td-total'>"
                         . "  <td class='align-left'>Latest Expenses Total</span></td>"
-                        . "  <td colspan='6' class='align-right'>" . number_format($total, 2, '.', ',') . "</td>"
+                        . "  <td colspan='7' class='align-right'>" . number_format($total, 2, '.', ',') . "</td>"
+                        . "  <td>&nbsp;</td>"
                         . "</tr>";
                         ?>
                     </tbody>
